로그인 되어있지않으면 글삭제 안되게 처리

+++ src/main/java/servlet/board/Remove.java
@@ -13,7 +13,7 @@ import dto.Criteria;
 import service.BoardService;
 import service.BoardServiceImpl;
 import utils.Commons;
-//import vo.Member;
+import vo.Member;
 
 @WebServlet("/board/remove")
 public class Remove extends HttpServlet{
@@ -21,20 +21,20 @@ public class Remove extends HttpServlet{
 	@Override
 	protected void service(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
 		String pnoStr = req.getParameter("pno");
-//		Object memberObject = req.getSession().getAttribute("member");
+		Object memberObject = req.getSession().getAttribute("member");
 		Criteria cri = new Criteria(req);
 		String redirectUrl = "list?" + cri.getQs2();
 		
-//		if(pnoStr == null || memberObject == null) {
-//			Commons.printMsg("비정상적인 접근입니다",redirectUrl, resp);
-//			return;
-//		}
+		if(pnoStr == null || memberObject == null) {
+			Commons.printMsg("비정상적인 접근입니다",redirectUrl, resp);
+			return;
+		}
 		Long pno = Long.valueOf(pnoStr);
-//		Member m = (Member) memberObject;
-//		if(!m.getId().equals(service.findBy(pno).getWriter())) {
-//			printMsg("본인이 작성한 글만 삭제할 수 있습니다.", redirectUrl, resp);
-//			return;
-//		}
+		Member m = (Member) memberObject;
+		if(!m.getId().equals(service.findBy(pno).getWriter())) {
+			Commons.printMsg("본인이 작성한 글만 삭제할 수 있습니다.", redirectUrl, resp);
+			return;
+		}
 		service.remove(pno);
